New model for interactive camera navigation
This replaces the previous implementation of the interactive
camera system that can be controlled by mouse/touch inputs
and on-screen controls. The reason is to fix some of the
seemingly inherent bugs with the previous implementation.
Ultimately, the previous system made more sense for a terrain
that is fully flat (i.e. the Y coordinate is always 0), but
didn't fully account for a terrain that is 3D.

The previous model was based around the camera orbiting a
point on the terrain. Zooming toward/rotating around a point
other than this center point required interpolating between
the orbit point and the new target point.

In the new model, the camera is more "free floating", and not
explictly tied to orbiting around a center point. Gestures
such as a zoom or rotation still have a central point that
they are oriented around, but this point only exists for the
duration of the gesture, and there is no concept of a permanent
center of orbit. The benefit is that this seems like a more
direct approach to model things, instead of "implement this in
terms of this other thing".

General highlights:

* MapCamera full replaces OrbitalCamera
* Seems conceptually simpler, and also requires less code
* Still has concept of "center of orbit", but now this point
  is determined in full 3D, instead of 2D with Y coordinate
  always 0.
* On desktop, zoom/rotation/tilt occur around terrain point at
  the center of the screen.
* Doesn't prevent intersection with terrain yet, so that is a
  reversion until that is implemented again.