World can be azimuth rotated with mouse when alt key is pressed
When the alt key and the main mouse button are pressed, moving
the mouse left/right causes the world to be rotated around the
initial mouse location, instead of panning the world.

This seems to expose that azimuth rotation is not working correctly
if the center of action is not the center of the screen, so that
should be fixed in a future commit.