World can be tilt rotated with mouse when alt key is pressed
When the alt key and the main mouse button are pressed, moving
the mouse up/down causes the world to be tilt rotated, instead
of panning the world. Since azimuth rotation is also tied to the
alt key, tilt rotation vs. azimuth rotation is chosen from whether
the amount of y movement is bigger (tilt rotation) or the amount
of x movement is bigger (azimuth rotation).

The switch between azimuth rotation and tilt rotation is not smooth,
because tilt rotation re-centers the center of action at the center
of the screen, while azimuth rotation does not. This should be
resolved in a future commit.