Camera prevented from going under terrain when zooming backwards
The camera is prevented from going under the terrain when zooming
forward, because the target of zoom is never blocked by any terrain
between it and the camera. However, when zooming backward this isn't
the case (for example, if you zoom out backwards into a mountain).

When this happens, the camera will be automatically raised to the
relevant terrain height. Arguably it would be better for the camera
to stop moving instead of being raised to match the terrain, but
it could be confusing why the camera stopped because you wouldn't
be able to see the obstruction.