All buildings rendered with a single geometry
This reduces the number of draw calls needed to render the buildings
from 50 down to 1. Quick and dirty performance measurements seem to
indicate that this reduces the amount of time needed to render the
entire scene. Although not a panacea, it seems to make rendering a
little smoother on a computer without a powerful graphic card.