Switching build process to Webpack, and adding Babel
This allows modularizing the code, which has several benefits. For
example, it will make it easier to limit the scope of variables
and objects without having to wrap them in a function.

This also removes the custom Ruby build script. This simplifies
the development environment by removing the need for Ruby, since
this was the only place it was used.

Although not actually being taken advantage of yet, incorporating
Babel will allow using more modern JavaScript features, and also
allow the tests to be ported to Jest in a future commit.

However, for the time being CSS is not being minified anymore,
because that hasn't been implemented via the new Webpack build
process yet. Also, the existing tests have not been modified and
do not work with the new build artifacts. That will be resolved in
a separate commit by switching the test runner to Jest.