Switching test runner from Jasmine to Jest
This is necessary to allow tests to be run against the artifacts
of the new Webpack build process. However, it also has the benefit
of allowing the tests to be run from the command line, and also
removes the need to manually update `SpecRunner.html` each time a
new test file is added.

A drawback is that the tests don't currently have access to
Three.js code. The one place that Three.js code was called in the
tests was `CityTourMath.randomInteger()`, which was a wrapper of
the `randInt()` Three.js function. To get the tests working,
`CityTourMath.randomInteger()` now contains the inlined
implementation of `randInt()`. However, in the future some way of
accessing Three.js in the tests should be implemented.

Also in the process of making these changes, a small bug has been
fixed in the tests for `SineEasing` (a test accidentally used
`LinearEasing` instead of `SineEasing`).