CSS minified in production Webpack builds
I currently don't understand why the `TerserPlugin` needs to be
defined in two separate places. If it's present in both the
`optimization` and `plugins` sections, then the minified
JavaScript in a production build is a little smaller (~2Kb) than
if it's missing from either place. Functionally the bigger and
smaller builds appear to be the same?