Mouse events work when mouse dragged outside canvas
Gestures started while the mouse is over the canvas, such as
rotating the world, zooming, or panning, all now continue to
respond to mouse movements after the mouse is dragged out of the
canvas (including out of the browser window itself). In addition,
releasing the mouse button while outside of the canvas will also
correctly end the appropriate action.

Previously, all actions would be halted when the mouse left the
canvas. This made gestures more difficult to perform.