Max zoom relative to city center, not target of zoom
When zooming out, the max distance you are allowed to zoom is now
defined by a max distance from the center of the city, rather than
a max distance from the target of the zoom movement. This
effectively means the max zoom is defined by a sphere surrounding
the city center.

This allows more sensible and predictable behavior when zooming
out a far distance. Previously, when zooming out the zoom could
sometimes stop seemingly prematurely even though it seemed like
(from a user perspective) there should be plently of room to keep
zooming out. As an example, suppose the camera is near the left
edge of the world and facing out to the void, the camera is tilted
almost level with the ground, and the target of the zoom movement
is a point on the left edge of the world. When zooming out
(backwards toward the right edge of the world), the zoom would
previously stop somewhere in the middle of the city, which seemed
odd.

This doesn't prevent the camera from moving out of this sphere
when other gestures occur. For example, this commit does not add
any limits on movement due to a pan gesture.