Preventing roads on edges of terrain
Roads on the very edge of the terrain aren't rendered correctly,
because half of the road hangs off of the terrain into the air.
There doesn't seem to be an obvious way to fix that without making
a larger change to how roads are implemented (since roads always
straddle the middle of a terrain tile), so this is a simple
workaround.