Better behavior when gesture target is in sky
Previously, when a new zoom, rotation, or tilt gesture was
started, a target point in the sky (instead of a point on the
terrain) could cause trouble. Since a target chosen from a point
in the sky can result in a target very far away it's easy to
quickly move away from the city and become lost and disoriented.

After this commit, if the target of action isn't a point on the
terrain, a ray will be cast from the camera's position, and the
target of zoom will be clamped to an intersecting point on an
invisible box surrounding the city. This prevents a gesture from
taking the camera far away from the city, while still allowing
gestures to work in a seemingly intuitive way (because the gesture
will still be based on where your fingers or mouse pointer were
when the gesture was started).

If the ray does _not_ intersect with the invisible box, then the
old behavior will be reverted to. That is, the target point won't
be clamped to the bounding box, and it will be possible to end up
lost in space. So, this commit does not completely solve the
problem, but seems like a step in the right direction.