Better implementation of hydraulic erosion
This commit fixes a couple of issues with the hydraulic erosion
implementation.

First, the max amount of soil present in the water is now
determined by a percentage of the water amount, rather than a raw
amount. This prevents situations where there is more dissolved
soil than there is water itself. A percentage also seems like a
more intuitive parameter to configure.

Second, when the final amount of water evaporates any soil that
was still in the water is deposited, instead of disappearing. This
means that when erosion is done, the world has the same amount of
soil as when erosion started. I.e., soil is never removed, it is
only moved around.

The biggest practical benefit of these changes is that the erosion
process more properly fills valleys with soil from higher terrain,
resulting in smooth flat ground. This results in terrain that is
both more realistic looking, but also more practical for building
cities on.