More flexible method of adding hills to terrain
The `addRandomPyramids()` function now allows defining a bounding
box into which pyramids will be placed, and no longer has special
logic to base the hill height on the position in the terrain.

The goal is to make this function more flexible and general
purpose to allow experimentation with different methods of hill
placement.