Water correctly shown on north/south side walls
Previously, columns on the north or south side walls would be
either solid blue if there was water in the column, and solid
green otherwise. Now each column will show land with water on top
of it (if present), to allow the side walls to look like a
cross-section.

This previously _was_ the case for the west/east side walls, but
after this commit it's the case for all walls. This was originally
only added for the west/east side walls because rivers currently
only exit on the west or east sides of the terrain. However while
experimenting with terrain generation options that can result in
water on any side of the terrain decided to fix this.

The original commit that added this functionality for the west
and east side walls was 9a40fd8b5011331c2908a643e9fc613fd4efb62a.