Deep water clamped to max side wall height
Previously if the water height at a place on the edge of the
terrain was deeper than the side wall height the rendered column
would extend down past the bottom of the side wall. No longer!

As an optimization this also avoids adding triangles for the
non-existant land in this case.