Adding grid lines underneath the terrain
This adds a grid of lines, resembling graph paper, under the
terrain in a square area a little bit more than 2x the terrain's
width and height.

The intention is to give some grounding to the world, so that it
feels less like it is randomly floating in space. As a first pass
I tried having the grid extend infinitely into the distance, but
this had the unintentional effect of making the city feel small in
comparison to the vast expanse that the infinite grid implied. The
finite grid seems to strike a balance between providing a sense of
grounding while not making the city seem tiny.