Removing unneeded configuration parameters
`centerX` and `centerZ` were always the same as the center
coordinates of the downtown neighborhood, so they were redundant.

I think the code is a little easier to understand with it more
explicit that the target end point of the road is the center of
the downtown neighborhood. (When I first read this code a minute
ago I thought it was saying the target was the center of the
terrain, not the center of the downtown neighborhood, until I read
more closely).