Adding function to flood terrain with water
That is, to cover all terrain less than a given height with water.
This is not a flood _fill_, that is, it applies the water height
across the entire terrain, instead of filling a particular pit
with water.

This is being added on an experimental basis to allow playing
around with new methods of generating terrain.