Preventing disconnected roads in road network
Previously it was possible in some situations for roads to be
added that weren't connected to the rest of the road network. This
was caused by not taking existing roads into account when building
a path from a neighborhood center to the rest of the city. This
could result in a path that included a virtual bridge that
conflicted with an existing actual bridge. When the path was then
converted into an actual road, building the new bridge would fail,
causing a disconnect in the road network.

This is fixed by taking the existing road network into account
when building a path to the city center, so that virtual bridges
are only added to the path if they could also be added in a real
road.