Removing ban on bridges bordering too much land
The original intention of this restriction was to prevent building
bridges where both ends were on the same side of a river, because
those often seem unrealistic. However, this didn't 100% prevent
that, and it seems like in the modern code there are probably
better ways to prevent that. (For example by adjusting edge
weights when choosing a candidate road path). Also, in certain
types of terrain it seems like this would prevent bridges that
otherwise seem reasonable, such as a bridge over shallow water
where one side is the ocean and the other side is a cliff face.