Removing configurable bridge building probability
This legacy condition was effectively disabled in the current
code, and doesn't really make sense with the current road building
logic. I.e., when connecting a neighborhood center to the rest of
the road network, there doesn't seem to be a reason why you would
want to randomly not build a bridge that would otherwise be needed
to reach the rest of the city.