Better location for generating neighborhoods debug mesh
The mesh that shows the locations of neighborhood centers for
debugging purposes doesn't have anything to do with the mesh for
buildings per se, so it seemed a little weird that they were
coupled together. This also allows `SceneView` to be a central
location for where to configure which debug meshes are enabled.