Better turns around corners in driving animation
Previously, the driving animation worked like this: the camera
"drove" to the next target point, reached the middle of the target
intersection, stopped, rotated to face the next target point, and
then continued driving. Obviously, that is not how driving works
in the real world.

Now when when making turns the camera will move in a smooth arc
toward the next target point. This is how an actual car would
drive, and looks a lot better.

This introduces a new `CurveAnimation` object which allows moving
along an arbitrary curve at a constant rate. Although initially
used to implement the new driving animation, the intention is that
it could also be used in other flythrough animations in the
future.

This implementation is more of a first pass though. There are
several known issues that should be fixed in the future:

* The camera jumps suddenly (instead of smoothly transitioning) in
  the following scenarios:
  * The transition from the initial "fly to city center" animation
    to the initial driving animation.
  * The transition from the "hovering" animation to the driving
    animation.
  * When performing a 180 degree U-turn
* When turning corners onto a bridge the camera can momentarily
  slide off the terrain toward the water (which fixes itself once
  the camera has moved onto the main bridge span).
* Sometimes the camera moves in a diagonal path (regardless of
  the presence of roads) for an unknown reason.