Adding ability to visualize curves for debugging
This allows an arbitrary set of Three.JS `Curve` objects to be
displayed in the world as a set of curved tubes. This is intended
to be used as a debugging aid for animations, such as the new
driving animation whose driving path is implemented as a Three.js
`Curve` object.