`getRoadHeight()` matches visual road appearance
Previously, when this function was called with X or Z coordinates
that weren't exactly on a road intersection, the road height was
interpolated between the height of the two intersections adjacent
to the relevant road edge. However, this did not match the visual
appearance of how roads are displayed in the world. A square area
around each intersection point is flat, and this needs to be taken
into account to get correct results.

This change is being made in advance of updating the driving
animation to use `getRoadHeight()` to compute its driving path,
and without this change it won't follow the height of roads
correctly.

The existing code that calls this function only calls it with
exact intersection coordinates, so no other code needs to be
modified to take the changed behavior into account.