Driving animation uses correct road heights
Previously, the driving animation followed road heights correctly
for the most part, but could "fall off" the road when making a
turn onto a bridge. This was because the driving logic did not
directly take the height of the road into account - instead it
relied on the fallback that keeps the camera a minimum height off
the ground. However, this failed when making a turn onto a bridge,
because the land next to a bridge entrance is usually a cliff.

Now each segment of the curve that defines the driving path
includes actual height components, based on the road heights at
the relevant locations. However, this requires making the curve
much more detailed than before. A curve segment now has to exist
for each road segment, and also for each road intersection. This
is necessary to make sure the height of the road is properly
followed while driving along a path. (If you do a simple
interpolation between the heights two road intersections, the
corresponding interpolated height will not match the stair-step
appearance of the actual road).

Due to the requirement for a more detailed animation curve, the
driving path generated by `RoadNavigator` is no longer simplified,
because that would prevent this more detailed curve from being
created.