Extracting function
The `buildDrivingAnimations()` function originally did two things.
First, it built an abstract driving path, and then secondly it
turned that abstract path into an animation curve. However, those
are two distinct pieces of logic, and extracting the animation
curve building logic to a separate function helps decompose this
large function into more manageable pieces.

Extracting the animation curve building logic also makes it easier
to use it in other places, such using it to add a mesh
representation of the curve to the world for debugging purposes.