Better variable names
The logic in this function involves building a set of curve
segments that connect a list of points. The task of building a
segment involves working with 3 consecutive points. "start",
"middle", and "end" seem like more clear names to use for these
points than "current", "next", and "subsequent". (They are all
"current" in a sense, so that term is potentially confusing, and
"subsequent" seems like it means the same thing as "next").