Consecutive driving animations don't cause glitches
Previously, when performing two driving animations in a row, the
camera would often go haywire, moving in a diagonal direction
regardless of roads or buildings. This was because a path segment
not in one of the cardinal directions was often added.

Now an arbitary number of consecutive driving animations can be
performed and the car will stay on the road. However, if the
ending angle of one animation is different from the starting angle
of the next, the camera will suddenly jump because no animation
between the two angles is performed. That is a target for a future
commit.