Fixing coordinates used for flythrough descent logic
Previously, when checking to see if there were roads at the target
points of the swoop down and initial driving segment, the
coordinates used were _absolute_ coordinates, instead of relative
offsets from the starting point. For example, if the starting
point was {-4, 12}, the old logic would check for a road at e.g.
{0,3}, instead of {-4, 12+3=15}. This could result in a swoop or
drive not being performed, even though it seemingly should have
been possible based on the roads present.