Sharing single vector for identical points
In a driving animation curve, the end point of one segment and the
start point of the next are always identical. This commit allows
a single `THREE.Vector3` to be used for both points. This makes
the code a little easier to maintain by reducing duplication.
Also, the reduction in the number of `THREE.Vector3` objects that
need to be created is a minor optimization.