Better point representation in `PathFinder` path array
Previously, `PathFinder.shortestPath()` returned a path in which
each point in the path was represented by a two element array. The
first element represented the X coordinate, and the second element
the Z coordinate.

Now, each point is represented by an object with named properties
for the X and Z coordinates. This seems more clear, because now
code using a path array can look like `path[i].z` (or whatever),
instead of `path[i][1]`.