Better camera height during turns in driving animation
The camera now more closely follows the road height when making
turns in the driving animation, especially when going up or down a
hill.

This is enabled by using a cubic Bezier curve to define the turn,
instead of a quadratic Bezier curve. The extra control point
allows defining the curve with more precision, to more closely
match the road height during the turn.

The turning radius is essentially the same as before; only the
Y coordinates of the curve should be affected.