Content has correct height in iPhone landscape mode
Previously, the bottom of the main canvas and the navigation panel
dipped below the visible portion of the screen when viewed in iOS
landscape orientation. This made things hard to use, since the
only way to get to the bottom of the page was to open a menu, and
swipe down on the menu body to scroll the page down. (Touching the
main canvas would not work, because that would instead result in a
pan gesture in the world).

The problem is that in iOS landscape orientation "100%", "100vh",
etc. correspond to the full _screen_ height, including the browser
chrome, instead of the visible content area height.

I'm not currently aware of a way to set the content height to be
the visible content area height using only CSS. However, this
height can be obtained via `window.innerHeight`, and using it to
set the height of the parent `<html>` element mostly fixes the
problem.

This commit does not fix the loading/error message overlay
content. Although no content will be hidden, the message will be
off center vertically due to the container element being taller
than the visible position of the page.