Rotation/tilt sliders work correctly with touch events
Previously, the rotation and tilt sliders did not set the center
of rotation/tilt properly when used via touch events. This
resulted in two problems. First, on initial page load, the sliders
would have no effect, due to the center of rotation/tilt being
undefined. Second, they would function after a different gesture
was first performed, but the center of rotation/tilt would be
re-used from the previous gesture, instead of being the center of
the screen.

This commit fixes the problem by using the existing event handlers
that set the center of rotation/tilt for mouse events for touch
events as well.