Fixing non-responsive touch events after window focus
On iOS, when a browser tab loses focus (e.g. from switching to a
different tab), and then regains focus, the "focus" event doesn't
seem to be fired on the `window` object until the user interacts
with one of the elements on the page. However, touch events on the
main canvas _don't_ seem to cause a "focus" event for some reason.

This resulted in a bug where after user came back to a tab with
the app running touch events would no longer seem to work on the
main canvas until a different element was interacted with. However,
touch events _were_ working, but no changes were being rendered
because the main app timer had not been re-enabled due to the lack
of a "focus" event. When the timer was finally reactivated, the
camera could suddenly jump to a new location to reflect any touch
events that had taken place.

This commit fixes the issue by forcing the main app timer to be
restarted any time the main canvas is touched.