Tilt gestures always relative to center of window
Previously, tilt gestures were performed relative to a "center of
action" based on the location of the mouse/touches when a zoom,
azimuth rotation, or tilt gesture was started. However, unless
this point was exactly at the center of the window, the camera
would jump when starting a tilt gesture, which was disorienting.

In practice, there doesn't seem to be an obvious reason why it
would ever be beneficial to start a tilt gesture relative to
anywhere other than the center of the window.