Zoom is proportional to mouse/touch movement
Previously, each zoom gesture moved the camera by the same amount
regardless of the distance traveled by the mouse, or the change in
distance between the touch points. This sometimes made zoom
gestures seem sluggish, and led to the unintuitive behavior that
many small mouse movements over a short distance would result in a
larger amount of zoom than a quick mouse movement over a large
distance. (And similarly for touch events).