Preventing camera jump after tilt->zoom->tilt
Previously, the camera position could jump at step #3 in the
following scenario:

1.) Start a single-touch tilt gesture (i.e. press mouse button
    and ALT key).
2.) While still pressing the mouse button, switch to a zoom
    gesture (i.e. release ALT key, and press SHIFT key).
3.) Switch back to a tilt gesture (i.e. release SHIFT key and
    press ALT key).

This only happened when the center of action was originally set
from a tilt gesture. I.e., this could be recreated by going
"no gesture->tilt->zoom->tilt", but not "no gesture->zoom->tilt".

The problem was that the center of tilt was not reset when
starting a zoom gesture. This meant that when starting a tilt
gesture after performing a zoom, the old center of tilt would be
used, instead of a center of tilt based on the new camera
position.

This was not a problem with azimuth rotation gestures, because
b6e7c6c274dd63338ce2ac8de022cbc7e8c587ac added the relevant logic
to handle this scenario for that gesture (but not zoom gestures).