Improving order of functions
The previous order of functions in the parent `GestureProcessor`
function was jumbled, and made the code annoying to work with.
This commit changes to order to have the functions generally go
"outside-in" as they go from top to bottom, which seems like it
will make things easier to work with.

No code is actually being changed, other than being moved around.