Normalizing generated terrain heights
This normalizes the height of generated terrain so that the
lowest point in the terrain always has a height of 0.0. (Land
height can go below 0.0, as long as the water surface height is
0.0).

This has a few benefits:

* Allows touch gestures to work correctly with parts of the
  terrain that (before normalization) dip below 0.0. Currently,
  the logic that maps screen coordinates to world coordinates
  clamps Y positions to 0.0 as a backstop, making this otherwise
  impossible.
* Prevents the background grid from covering parts of the terrain
  that dip below the background grid's Y position.
* Ensures sidewalls always have a minimum height, regardless of
  the height of the terrain.

Alternately, this could be solved by keeping terrain heights
as-is but making other components/logic more flexible. However,
normalizing seems simpler to implement, makes the world data
easier to reason about, and it's unclear what specific benefits
added flexibility would have in the future.

Normalization seems to take 1-3ms on my laptop. While not 0, this
is neglible compared to the rest of the time needed to generate a
new city.