Camera matrix updated after pan or zoom gesture
After a `THREE.Camera` position is changed, its corresponding
world matrix must be updated for screen->world coordinate
calculations to be correct. Normally, this gets updated
automatically by the `THREE` scene renderer each time a render
occurs. However, if multiple gesture events happen between two
renders, this can result in the gestures after the first to be
based on incorrectly calculated world positions, unless this is
handled somehow.

This commit ensures that the relevant matrix is always updated
after a pan or zoom gesture, to avoid this problem. This was
already the case for tilt and azimuth rotation gestures, because
this problem was exposed earlier for them due to how quickly they
can be switched between with single touch gestures.

I'm not sure what practical effect this commit has, if any,
because I don't see any obvious difference in panning or zooming
behavior as a result. I only realized the need to update the
matrices after noticing a debug marker for the mouse's world
position jumped around unexpectedly during a pan gesture. However,
it seems like the world position should be made to be correct, for
basic hygiene at a minimum.