Fixing tilt rate for single touch gestures
This changes the tilt rate so that a full tilt from the minimum
tilt angle to the maximum tilt (or vice versa) will occur when the
mouse/touch is dragged the full height of the browser window. I'm
not sure exactly what the previous required drag distance of a
full tilt was, but it was somewhat larger than the window height,
and felt arbitrary.

This makes the single touch tilt gesture more rationally unified
with the multi-touch tilt gesture, which performs a full tilt
after moving the touches half the height of the browser window.
(This difference in tilt rate is a result of the single touch rate
feeling better on a desktop/laptop computer, and the multi-touch
rate feeling better on a phone, in my subjective judgement).