Resetting center of action upon pan gesture
It's possible to switch between the 4 main gestures (pan, zoom,
azimuth rotation, tilt rotation) without releasing a mouse button
or finger. Previously, when switching between gestures in this
manner, the following scenario could cause a problem:

1.) Start a zoom/azimuth rotate/tilt gesture, which will set the
    center of action based on the mouse/touch positions.
2.) Switch to a pan gesture without releasing the mouse button or
    final touch point.
3.) Switch to a zoom/azimuth rotate/tilt gesture.

Previously, at step #3, the center of action would not have been
reset. Since the mouse or touches could be in a different place
than the original location in step #1, this would cause gestures
to occur relative to a possibly unexpected center point.

Now, the center of action will be set to `undefined` at each time
a pan gesture occurs (i.e. step #2 above). At step #3, this will
cause a new center of action to be set from the new current mouse
or touch location.