Touching screen during flythrough locks camera angle
Previously, touching the screen during a flythrough would cause
the camera to be rotated a relative amount from the vehicle's
base rotation around the X and Y axes. This allowed you to look
around the world during a flythrough, instead of always looking
straight ahead. However, if you were turned to look at something,
and the vehicle turned, the camera would turn with it, causing
your view to be lost.

Now, when touching the screen, the camera will be locked to look
at the current direction, regardless of the vehicle's future
rotation. As the vehicle turns, you will still look in the same
original direction. This makes it easier to follow something in
the world (a hill, a bridge, downtown, etc.) as the vehicle moves
around.

This change was motivated by continual annoyance at losing sight
of something interesting in the world during a flythrough because
the vehicle turned. After trying this change out, it seems like an
improvement.