Initial height of swoop down takes terrain into account
Previously, the initial climb at the beginning of a flythrough
reached a constant absolute height, regardless of the height of
the underlaying terrain. This meant that if the city center was on
higher ground, the initial swoop down would be shorter and less
dramatic than if the city was on lower ground.

Now the initial climb will reach a height relative to the height
of the undelaying terrain, so the height of the swoop down should
be the same for any city.