Preventing camera jump when locking camera angles
Previously, if a "reset to center" animation was in progress
during a flythrough, touching the screen to lock the angle would
not take the animation into account, potentially causing the
camera to jump to an unexpected location.

Now the angle will be locked to whatever the current visible angle
is at the time of the touch event.