Tilt angle animated after flythrough
During a flythrough, the camera is allowed to be tilted more level
with the horizon than during the normal "interactive map mode".
Previously, when a flythrough was stopped, the tilt angle was not
adjusted to the minimum "map mode" tilt angle. The next
interaction with the world would then cause the camera to jump to
the minimum tilt angle.

Now the camera will smoothly animate from the final flythrough
tilt angle to the closest valid tilt angle in "map mode" (if
needed).