Better panning speed
Previously, when the camera was tilted to be facing close to
straight at the horizon, panning the world was way too slow,
especially if you were zoomed closed to the terrain.

Now, panning when the camera is tilted close to facing the horizon
is much faster, which in practice makes it much more usable and
pleasant.

This conceptually uses the same "always pan as if you were looking
straight down" logic as before (although implemented in a more
clear way). The reason the panning speed is faster is that it
bases the "looking down" position on where you would be if you
performed a normal tilt gesture (which will raise you higher into
the sky), rather than the actual camera position. The higher the
vantage point, the faster the pan.

One thing to note is that the panning speed can change while
panning around, depending on what terrain is at the center of the
window at any given moment. (I.e., what point in the world will be
used as the "center of tilt"). This effect is more pronounced the
more the camera is tilted to look straight at the horizon. This
seems a little weird and ideally should be fixed in the future,
but for now this still seems like an improvement over before.