Downtown neighborhood debug marker highlighted
The marker for the neighborhood that represents downtown is now
colored yellow, instead of magenta like the other markers. This
makes it easier to tell at a glance which neighborhood represents
downtown.

A side consequence is that this allows removing the legacy debug
marker in `SceneView` that duplicated the same functionality (but
in a less useful way).