Correct center of action on grid plane, sidewalls
Previously, the center of action for gestures would not be set
correctly if the target was on one of the terrain sidewalls, or
the grid plane.

If the camera->world position ray intersected the invisible box
surrounding the city, but the intersection Y coordinate was below
0.0 (either on the lower part of a sidewall, or on the grid
plane), the center of action would be set to an underground point
on the _opposite_ side of the city. This made rotations confusing,
and allowed zooming such that you could see the underside of the
city. (At least, until the camera was inside the terrain bounds,
in which case it would suddenly jump up to terrain height).

If the camera->world position ray did not intersect the invisible
box, the target Y position would be clamped to 0.0, instead of the
top of the grid plane (which is lower than 0.0). While a less
serious issue, this meant that the center of action did not match
where it should be visually.