Preventing roads and bridges on terrain edges
This is a follow on from 56ba22c05908998d44110119cdb6ef5a961ba791,
which prevented roads inside neighborhoods from being directly on
the terrain edges. This new commit also prevents roads (including
bridges) that connect neighborhoods from being on the terrain
edges as well.

As noted in the original commit, roads on the edge of the terrain
look weird because half of the road surface hangs off of the
terrain into the air.

This commit introduces fields to `RoadNetwork` that indicate the
bounding box that all roads must be inside. This allows
centralizing the relevant logic, removing the need to re-implement
it each place it is needed.

This also requires updating some test fixture data to compensate
for some roads in tests that previously were on the edges of the
terrain.