More specific function names
The word "bounding" hopefully makes it more clear that these refer
to the bounding box around the existing nodes, especially now that
there are also similar fields for the allowed range that roads can
be added in.

The use of "X" and "Z" instead of "column" and "row" are intended
to help standardize terminology across the app. "X" and "Z" seem
more explicit, and don't require a mental translation to figure
out what dimensions "column" and "row" refer to.