Driving U-turn doesn't cause camera jump
Previously, when the driving animation required the "car" to
perform a U-turn, the camera would instantly jump from the old
"forward" angle to the new forward angle (flipping 180 degrees).
This was jarring and disorienting.

Now when a U-turn is required the camera will smoothly perform a
180 degree rotation (in place) to the new driving direction before
continuing on its way.