Initial driving animation doesn't go off road
Previously, when determining the path of the initial 3 block
driving segment at the start of a flythrough, the only check made
for a valid path was that the target road intersection existed.
However, if road segments were missing along this path it would
cause the camera to go off road.

Now the initial driving target will only be chosen such that there
is a straight line path to it that doesn't require going off road.
If there is not a valid path 3 blocks long, a valid 2 block, 1
block, or 0 block path will be chosen instead.